DS-146: Create breakpoint-utils, move utility constants related to BreakpointObserver here
