+++ ngx-fudis/projects/ngx-fudis/src/lib/services/grid/grid.service.ts
@@ -2,7 +2,7 @@ import { BreakpointObserver, BreakpointState } from '@angular/cdk/layout';
 import { Injectable, Signal, signal } from '@angular/core';
 
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
-import { breakpointsMinWidthToObserve } from '../../directives/grid/gridUtils';
+import { breakpointsMinWidthToObserve } from '../../directives/breakpoint/breakpoint-utils';
 import { FudisGridAttributes, FudisGridResponsiveData } from '../../types/grid';
 
 @Injectable()
