+++ ngx-fudis/projects/ngx-fudis/src/lib/services/spacing/spacing.service.ts
@@ -2,7 +2,7 @@ import { BreakpointObserver, BreakpointState } from '@angular/cdk/layout';
 import { Injectable, signal } from '@angular/core';
 
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
-import { breakpointsMinWidthToObserve } from '../../directives/spacing/spacing-utils';
+import { breakpointsMinWidthToObserve } from '../../directives/breakpoint/breakpoint-utils';
 import { FudisSpacingResponsiveData } from '../../types/spacing';
 
 @Injectable()
